# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "FeedMeta": {
        "properties": {
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "FeedMeta",
        "type": "object"
      },
      "FeedResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/FeedDocument"
            },
            "title": "Documents",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/FeedMeta"
          }
        },
        "required": [
          "meta",
          "documents"
        ],
        "title": "FeedResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LiteratureIngestRequest": {
        "properties": {
          "pmids": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Pmids",
            "type": "array"
          }
        },
        "required": [
          "pmids"
        ],
        "title": "LiteratureIngestRequest",
        "type": "object"
      },
      "LiteratureRetrieveRequest": {
        "properties": {
          "date_range": {
            "default": "any",
            "title": "Date Range",
            "type": "string"
          },
          "max_results": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Results",
            "type": "integer"
          },
          "query": {
            "maxLength": 400,
            "minLength": 2,
            "title": "Query",
            "type": "string"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          },
          "study_types": {
            "items": {
              "type": "string"
            },
            "title": "Study Types",
            "type": "array"
          }
        },
        "required": [
          "query"
        ],
        "title": "LiteratureRetrieveRequest",
        "type": "object"
      },
      "LiteratureSearchRequest": {
        "properties": {
          "expand": {
            "default": false,
            "title": "Expand",
            "type": "boolean"
          },
          "max_results": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Results",
            "type": "integer"
          },
          "query": {
            "maxLength": 400,
            "minLength": 2,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "LiteratureSearchRequest",
        "type": "object"
      },
      "LiteratureSearchResult": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/FeedDocument"
            },
            "title": "Documents",
            "type": "array"
          },
          "effective_query": {
            "title": "Effective Query",
            "type": "string"
          },
          "embedded": {
            "title": "Embedded",
            "type": "boolean"
          },
          "found": {
            "title": "Found",
            "type": "integer"
          },
          "ingested": {
            "title": "Ingested",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "effective_query",
          "found",
          "ingested",
          "embedded",
          "documents"
        ],
        "title": "LiteratureSearchResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "augura-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/corpus/feed": {
      "get": {
        "operationId": "feed_corpus_feed_get",
        "parameters": [
          {
            "in": "query",
            "name": "jurisdiction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jurisdiction"
            }
          },
          {
            "in": "query",
            "name": "evidence_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Evidence Type"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feed",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/literature": {
      "post": {
        "description": "Agent de recherche de littérature : cherche sur PubMed (E-utilities NCBI) et\ningère les articles dans le corpus du tenant (Document + Chunk). Embedder/LLM\noptionnels (sans clé : ingestion sans vecteur, requête non élargie).",
        "operationId": "literature_corpus_literature_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiteratureSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiteratureSearchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Literature",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/literature/ingest": {
      "post": {
        "description": "Ingère des enregistrements PubMed précis (par PMID) dans le corpus du tenant.\nSert « Add to corpus » sur des résultats de retrieve gardés (PubMed only).",
        "operationId": "literature_ingest_corpus_literature_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiteratureIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiteratureSearchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Literature Ingest",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/literature/retrieve": {
      "post": {
        "description": "Récupère en direct (PubMed + CT.gov), groupé par source, streamé en NDJSON.\nNe touche PAS au corpus (aucune ingestion). known-item ⇒ source unique ; topique\n⇒ fan-out parallèle. Le query_string exact par résultat est porté pour le gel.",
        "operationId": "literature_retrieve_corpus_literature_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiteratureRetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Literature Retrieve",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/literature/sessions": {
      "delete": {
        "description": "Vide l'historique « Recent queries » du tenant (nettoyage en un clic).",
        "operationId": "clear_sessions_corpus_literature_sessions_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Clear Sessions",
        "tags": [
          "corpus"
        ]
      },
      "get": {
        "operationId": "list_sessions_corpus_literature_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchSession"
                  },
                  "title": "Response List Sessions Corpus Literature Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "corpus"
        ]
      },
      "post": {
        "operationId": "create_session_corpus_literature_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/literature/sessions/{session_id}": {
      "delete": {
        "description": "Supprime une entrée de l'historique « Recent queries » (suppression unitaire).",
        "operationId": "delete_session_corpus_literature_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "corpus"
        ]
      },
      "get": {
        "operationId": "get_session_corpus_literature_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/literature/sessions/{session_id}/events": {
      "post": {
        "operationId": "append_event_corpus_literature_sessions__session_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventAppendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiteratureEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Event",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/literature/snapshots": {
      "get": {
        "description": "Liste les preuves gelées du tenant (vue légère, sans results ni recalcul de hash).",
        "operationId": "list_snapshots_corpus_literature_snapshots_get",
        "parameters": [
          {
            "in": "query",
            "name": "study_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SnapshotSummary"
                  },
                  "title": "Response List Snapshots Corpus Literature Snapshots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Snapshots",
        "tags": [
          "corpus"
        ]
      },
      "post": {
        "description": "Gèle le jeu de résultats + annotations : calcule le content_hash, épingle\nmodel/prompt version, persiste. La réponse porte le hash (auto-vérifiable).",
        "operationId": "create_snapshot_corpus_literature_snapshots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiteratureSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Snapshot",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/literature/snapshots/{snapshot_id}": {
      "get": {
        "description": "Relit un snapshot et VÉRIFIE le content_hash (erreur dure si divergence).\nLecture base pure : zéro appel PubMed/CT.gov (rejeu reproductible).",
        "operationId": "read_snapshot_corpus_literature_snapshots__snapshot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiteratureSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Snapshot",
        "tags": [
          "corpus"
        ]
      }
    },
    "/corpus/search": {
      "post": {
        "operationId": "search_corpus_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchHit"
                  },
                  "title": "Response Search Corpus Search Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "corpus"
        ]
      }
    },
...
  }
}